FW-4257 optimize my-sites queryset
